Data.HFunctor.Foldable

Documentation

type HFunction i a b = forall ix. i ix -> a ix -> b ix

Indexed, higher-kinded function

o :: HFunction i b c -> HFunction i a b -> HFunction i a c infixr 9

Composition of HFunctions Don't want to create a newtype wrapper for a Category instance because it makes it annoying to write HFunctions by hand

identity :: HFunction i a a

Typeclasses

class HFunctor i h where

A covariant functor over the category of kinds `Type -> Type`, where the domain is determined by the index i, with the arrows as functions

Methods

hmap :: HFunction i a b -> HFunction i (h a) (h b)

class HFunctor i h => HComonad i h where

HComonad

Minimal complete definition

hextract, (hduplicate | hextend)

Methods

hextract :: HFunction i (h a) a

hduplicate :: HFunction i (h a) (h (h a))

hextend :: HFunction i (h a) b -> HFunction i (h a) (h b)

class HFunctor i h => HMonad i h where

HMonad

Minimal complete definition

hpure, (hjoin | hbind)

Methods

hpure :: HFunction i a (h a)

hjoin :: HFunction i (h (h a)) (h a)

hbind :: HFunction i a (h b) -> HFunction i (h a) (h b)

class HFunctor i h => HRecursive i h where

An HFunctor that can be projected into from an index

Methods

hproject :: HFunction i Identity (h Identity)

class HFunctor i h => HCorecursive i h where

An HFunctor that can be embedded into an index

Methods

hembed :: HFunction i (h Identity) Identity

Destructive Morphisms

cata :: forall i h a. HRecursive i h => HFunction i (h a) a -> HFunction i Identity a

Catamorphism: ∀a. (f a → a) → μf → a Deconstructs an f-structure level-by-level and applies the algebra

para :: forall i h a. HRecursive i h => HFunction i (h (Product Identity a)) a -> HFunction i Identity a

Paramorphism: ∀a. (f (μf * a) → a) → μf → a Like cata, but allows access to the full subtree during teardown

zygo :: HRecursive i h => HFunction i (h b) b -> HFunction i (h (Product b a)) a -> HFunction i Identity a

Zygomorphism: ∀a b. (f b → b) → (f (b * a) → a) → μf → a Allows depending on a helper algebra for deconstructing an f-structure

histo :: HRecursive i h => HFunction i (h (HCofree h a)) a -> HFunction i Identity a

Histomorphism: ∀a. (f (Cofree f a) → a) → μf → a

prepro :: forall i h h' a. (HRecursive i h, HCorecursive i h') => (forall b. HFunction i (h b) (h' b)) -> HFunction i (h a) a -> HFunction i Identity a

Constructive Morphisms

ana :: forall i h a. HCorecursive i h => HFunction i a (h a) -> HFunction i a Identity

Anamorphism: ∀a. (a → f a) → a → νf

apo :: forall i h a. HCorecursive i h => HFunction i a (h (Sum Identity a)) -> HFunction i a Identity

Apomorphism: ∀a. (a → f (νf + a)) → a → νf

gapo :: HCorecursive i h => HFunction i b (h b) -> HFunction i a (h (Sum b a)) -> HFunction i a Identity

GApomorphism: ∀a b. (b → f b) → (a → f (b + a)) → a → νf

futu :: HCorecursive i h => HFunction i a (h (HFree h a)) -> HFunction i a Identity

Futumorphism: ∀a. (a → f (Free f a)) → a → νf

postpro :: forall i h h' a. (HCorecursive i h', HRecursive i h) => (forall b. HFunction i (h b) (h' b)) -> HFunction i a (h' a) -> HFunction i a Identity

Postpromorphism: ∀a. (a → f a) → (f ~> f) → a → νf

Combined Morphisms

hylo :: forall i h a b. HFunctor i h => HFunction i (h b) b -> HFunction i a (h a) -> HFunction i a b

Hylomorphism: ∀a b. (f b → b) → (a → f a) → a → b

dyna :: forall i h a b. HFunctor i h => HFunction i (h (HCofree h b)) b -> HFunction i a (h a) -> HFunction i a b

Dynamorphism: ∀a b. (f (Cofree f b) → b) → (a → f a) → a → b

chrono :: HFunctor i h => HFunction i (h (HCofree h b)) b -> HFunction i a (h (HFree h a)) -> HFunction i a b

Chronomorphism: ∀a. (f (Cofree f b) → b) → (a → f (Free f a)) → a → b

Constructive Distribution Laws

type Law i f g = forall a. HFunction i (f (g a)) (g (f a))

An HFunction to distribute HFunctor f over HFunctor g

distCata :: HFunctor i h => Law i h Identity1

Distribution law for cata. Note the Identity1 constructor is removed in cata.

distPara :: HCorecursive i h => Law i h (Product Identity)

Distribution law for para.

distZygo :: HFunctor i h => HFunction i (h b) b -> Law i h (Product b)

Distribution law for zygo.

distHisto :: HFunctor i h => Law i h (HCofree h)

Distribution law for histo

Destructive Distribution Laws

distAna :: HFunctor i h => Law i Identity1 h

Distribution law for ana

distApo :: HRecursive i h => Law i (Sum Identity) h

Distribution law for apo

distGApo :: HFunctor i h => HFunction i b (h b) -> Law i (Sum b) h

Distribution law for gapo

distFutu :: HFunctor i h => Law i (HFree h) h

Distribution law for futu

Generics

gfold :: forall i h w a. (HRecursive i h, HComonad i w) => Law i h w -> HFunction i (h (w a)) a -> HFunction i Identity a

Generic fold

gunfold :: forall i h m a. (HCorecursive i h, HMonad i m) => Law i m h -> HFunction i a (h (m a)) -> HFunction i a Identity

Generic unfold

ghylo :: forall i w f m a b. (HComonad i w, HFunctor i f, HMonad i m) => Law i f w -> Law i m f -> HFunction i (f (w b)) b -> HFunction i a (f (m a)) -> HFunction i a b

Generic hylomorphism, fold + unfold

Useful Types

newtype Identity1 f ix

Identity HFunctor

Constructors

Identity1

Fields

runIdentity1 :: f ix

data HCofree h a ix

Higher-kinded Cofree Comonad

Constructors

(a ix) :< (h (HCofree h a) ix) infixr 5

data HFree h a ix

Higher-kinded Free Monad

Constructors

HPure (a ix)

HFree (h (HFree h a) ix)