Data.IFunctor.Foldable

Fixpoint for indexed types

newtype IFix f ix

Fixpoint type

Constructors

IFix

Fields

unIFix :: f (IFix f) ix

Morphisms

Constructive morphisms

cata :: IFunctor f => (f a ~~> a) -> IFix f ~~> a

Catamorphism, ie. fold

prepro :: IFunctor f => (forall b. f b ~~> f b) -> (f a ~~> a) -> IFix f ~~> a

Fokkinga's prepromorphism

para :: IFunctor f => (f (Product (IFix f) a) ~~> a) -> IFix f ~~> a

Paramorphism

zygo :: IFunctor f => (f b ~~> b) -> (f (Product b a) ~~> a) -> IFix f ~~> a

Zygomorphism

histo :: IFunctor f => (f (ICofree f a) ~~> a) -> IFix f ~~> a

Histomorphism

Destructive morphisms

ana :: IFunctor f => (a ~~> f a) -> a ~~> IFix f

Anamorphism, ie. unfold

postpro :: IFunctor f => (forall b. f b ~~> f b) -> (a ~~> f a) -> a ~~> IFix f

Fokkinga's postpromorphism

apo :: IFunctor f => (a ~~> f (Sum (IFix f) a)) -> a ~~> IFix f

Apomorphism

gapo :: IFunctor f => (b ~~> f b) -> (a ~~> f (Sum b a)) -> a ~~> IFix f

GApomorphism

futu :: IFunctor f => (a ~~> f (IFree f a)) -> a ~~> IFix f

Futumorphism

Combined morphisms

hylo :: IFunctor f => (f b ~~> b) -> (a ~~> f a) -> a ~~> b

Hylomorphism, fold then unfold

dyna :: IFunctor f => (f (ICofree f b) ~~> b) -> (a ~~> f a) -> a ~~> b

Dynamorphism

chrono :: IFunctor f => (f (ICofree f b) ~~> b) -> (a ~~> f (IFree f a)) -> a ~~> b

Chronomorphism

meta :: (IFunctor f, IFunctor g) => (f a ~~> a) -> (a ~~> b) -> (b ~~> g b) -> IFix f ~~> IFix g

Metamorphism TODO: ensure only 1 pass over structure

elgot :: forall f a b. IFunctor f => (f a ~~> a) -> (b ~~> Sum a (f b)) -> b ~~> a

coelgot :: forall f a b. IFunctor f => (Product a (f b) ~~> b) -> (a ~~> f a) -> a ~~> b

Monadic morphisms

cataM :: (ITraversable f, Monad m) => (forall ix. SingI ix => f a ix -> m (a ix)) -> forall ix. SingI ix => IFix f ix -> m (a ix)

Monadic catamorphism

paraM :: (ITraversable f, Monad m) => (forall ix. SingI ix => f (Product (IFix f) a) ix -> m (a ix)) -> forall ix. SingI ix => IFix f ix -> m (a ix)

Monadic Paramorphism

anaM :: (ITraversable f, Monad m) => (forall ix. SingI ix => a ix -> m (f a ix)) -> forall ix. SingI ix => a ix -> m (IFix f ix)

Monadic anamorphism

apoM :: (ITraversable f, Monad m) => (forall ix. SingI ix => a ix -> m (f (Sum (IFix f) a) ix)) -> forall ix. SingI ix => a ix -> m (IFix f ix)

Monadic apomorphism

hyloM :: (ITraversable f, Monad m) => (forall ix. SingI ix => f b ix -> m (b ix)) -> (forall ix. SingI ix => a ix -> m (f a ix)) -> forall ix. SingI ix => a ix -> m (b ix)

Monadic hylomorphism

dynaM :: (ITraversable f, Monad m) => (forall ix. SingI ix => f (ICofree f b) ix -> m (b ix)) -> (forall ix. SingI ix => a ix -> m (f a ix)) -> forall ix. SingI ix => a ix -> m (b ix)

Monadic dynamorphism

Distribution Laws

type DistLaw f g = forall a. f (g a) ~~> g (f a)

Type of distribution laws A Constructive morphism means the second part is a IComonad A Destructive morphism means the first part is an IMonad

Constructive distribution laws

distCata :: IFunctor f => DistLaw f IIdentity

distPara :: IFunctor f => DistLaw f (Product (IFix f))

distZygo :: IFunctor f => (f b ~~> b) -> DistLaw f (Product b)

distHisto :: IFunctor f => DistLaw f (ICofree f)

Destructive distribution laws

distAna :: IFunctor f => DistLaw IIdentity f

distApo :: IFunctor f => DistLaw (Sum (IFix f)) f

distGApo :: IFunctor f => (b ~~> f b) -> DistLaw (Sum b) f

distFutu :: IFunctor f => DistLaw (IFree f) f

Generalized morphisms

gfold :: forall f w a. (IFunctor f, IComonad w) => DistLaw f w -> (f (w a) ~~> a) -> IFix f ~~> a

Generalized fold

gunfold :: forall f m a. (IFunctor f, IMonad m) => DistLaw m f -> (a ~~> f (m a)) -> a ~~> IFix f

Generalized unfold

ghylo :: forall f w m a b. (IFunctor f, IComonad w, IMonad m) => DistLaw f w -> DistLaw m f -> (f (w b) ~~> b) -> (a ~~> f (m a)) -> a ~~> b

Generalized hylomorphism

gprepro :: forall f w a. (IFunctor f, IComonad w) => DistLaw f w -> (forall c. f c ~~> f c) -> (f (w a) ~~> a) -> IFix f ~~> a

Generalized prepromorphism

gpostpro :: forall f m a. (IFunctor f, IMonad m) => DistLaw m f -> (forall c. f c ~~> f c) -> (a ~~> f (m a)) -> a ~~> IFix f

Generalized postpromorphism

Generalized monadic morphisms

gfoldM :: forall f w m a. (ITraversable f, ITraversable w, IComonad w, Monad m) => DistLaw f w -> (forall ix. SingI ix => f (w a) ix -> m (a ix)) -> forall ix. SingI ix => IFix f ix -> m (a ix)

Generalized monadic fold

gunfoldM :: forall f m a x. (ITraversable f, ITraversable m, IMonad m, Monad x) => DistLaw m f -> (forall ix. SingI ix => a ix -> x (f (m a) ix)) -> forall ix. SingI ix => a ix -> x (IFix f ix)

Generalized monadic unfold

ghyloM :: forall f w m a b x. (ITraversable f, ITraversable w, ITraversable m, IComonad w, IMonad m, Monad x) => DistLaw f w -> DistLaw m f -> (forall ix. SingI ix => f (w b) ix -> x (b ix)) -> (forall ix. SingI ix => a ix -> x (f (m a) ix)) -> forall ix. SingI ix => a ix -> x (b ix)

Generalized monadic hylomorphism

Re-exports

newtype Const a (b :: k)

The Const functor.

Constructors

Const

Fields

getConst :: a

data Product (f :: k -> Type) (g :: k -> Type) (a :: k)

Lifted product of functors.

Constructors

Pair (f a) (g a)

data Sum (f :: k -> Type) (g :: k -> Type) (a :: k)

Lifted sum of functors.

Constructors

InL (f a)

InR (g a)

class IFunctor f => IComonad f where

Comonoid in the category of dependent endofunctors

Minimal complete definition

iextract, (iduplicate | iextend)

Methods

iextract :: f a ~~> a

iduplicate :: f a ~~> f (f a)

iextend :: (f a ~~> b) -> f a ~~> f b

type (~~>) a b = forall ix. SingI ix => a ix -> b ix infixr 4

Indexed function type

class IFunctor f where

Functor in the category of dependent types

Methods

imap :: (a ~~> b) -> f a ~~> f b

module Data.IFunctor.Classes

data ICofree f a ix

Cofree IComonad

Constructors

(a ix) ::< (f (ICofree f a) ix) infixr 5

data IFree f a ix

Free IMonad

Constructors

IPure (a ix)

IFree (f (IFree f a) ix)

data IIdentity f ix

Constructors

IIdentity

Fields

runIIdentity :: f ix

class IFunctor f => IMonad f where

Minimal complete definition

ipure, (ijoin | ibind)

Methods

ipure :: a ~~> f a

ijoin :: f (f a) ~~> f a

ibind :: (a ~~> f b) -> f a ~~> f b

class IFunctor f => ITraversable f where

Methods

itraverse :: (Applicative m, SingI ix) => (forall ix. SingI ix => a ix -> m (b ix)) -> f a ix -> m (f b ix)

imapDefault :: ITraversable f => (a ~~> b) -> f a ~~> f b

Default imap for deriving IFunctor

class SingI (a :: k) where

Methods

sing :: Sing a